+++ akg_agents/benchmark/kernelbench/mindspore/13_matmul_for_symmetric_matrices/13_matmul_for_symmetric_matrices_mindspore.py
@@ -31,9 +31,9 @@ N = 4096
 
 def get_inputs():
     
-    A = ms.mint.randn(N, N, dtype=ms.float16)
+    A = ms.mint.randn(N, N, dtype=ms.float32)
     A = (A + A.T) / 2  # Ensure symmetry
-    B = ms.mint.randn(N, N, dtype=ms.float16)
+    B = ms.mint.randn(N, N, dtype=ms.float32)
     B = (B + B.T) / 2  # Ensure symmetry
     return [A, B]
