+++ akg_agents/benchmark/kernelbench/mindspore/6_matmul_with_large_k_dimension_/6_matmul_with_large_k_dimension__mindspore.py
@@ -32,8 +32,8 @@ K = 131072
 
 
 def get_inputs():
-    A = ms.mint.randn(M, K, dtype=ms.float16)
-    B = ms.mint.randn(K, N, dtype=ms.float16)
+    A = ms.mint.randn(M, K, dtype=ms.float32)
+    B = ms.mint.randn(K, N, dtype=ms.float32)
     return [A, B]
